+++ src/main/kotlin/dev/mizarc/waystonewarps/application/actions/world/IsValidWarpBase.kt
@@ -1,0 +1,9 @@
+package dev.mizarc.waystonewarps.application.actions.world
+
+import dev.mizarc.waystonewarps.application.services.ConfigService
+
+class IsValidWarpBase(private val configService: ConfigService) {
+    fun execute(blockName: String): Boolean {
+        return configService.getStructureBlocks(blockName).isNotEmpty()
+    }
+}