+++ src/main/kotlin/dev/mizarc/waystonewarps/application/results/UpdateWarpSkinResult.kt
@@ -1,0 +1,8 @@
+package dev.mizarc.waystonewarps.application.results
+
+enum class UpdateWarpSkinResult {
+    SUCCESS,
+    WARP_NOT_FOUND,
+    BLOCK_NOT_VALID,
+    UNCHANGED
+}